Major: delegate cameras to a future branch
This is the last attempt to support security cameras in master branch.

A new branch will be created, to be re-opened in the future, possibly.

IP Camera support should be able to handle situations where the image is
actually an mjpeg, there's no Content-Length and HTTP Chunked Transfer
Encoding is in use.

See also: http://forum.xbmc.org/showpost.php?p=979387&postcount=2

Moreover, proper support for video streming, rtsp etc. is needed, and
development is not even started right now.